refactor: centralize auth boot wiring in createAuth
Fire onLogin at construction time if a cached session exists in the
store. This eliminates the orphaned boot code from honeycrisp,
opensidian, and zhongwen—encryption keys are now applied in one place.

Tab-manager keeps its async boot code because chrome.storage is async
(session.current starts as null until it loads). createAuth's
construction-time check is a no-op for async stores.

Also: rename authSession → session in tab-manager and move auth.ts
from $lib/state/ to $lib/ to match the other three apps.
Deduplicate AuthenticatedSession construction in BA subscription.
